Move the tests into individual files for parallelization
