+++ packages/fast-cli/src/cli.spec.ts
@@ -211,307 +211,4 @@ test.describe("CLI", () => {
             ).not.toThrow();
         });
     });
-    test.describe("add-foundation-component", () => {
-        test.describe("blank", () => {
-            test.beforeAll(() => {
-                setup(tempDir, tempComponentDir);
-                execSync(`cd ${tempDir} && npm run fast:init`);
-                setup(tempDir, tempComponentDir);
-                execSync(`cd ${tempDir} && npm run fast:add-foundation-component:blank`);
-            });
-            test.afterAll(() => {
-                teardown(tempDir, tempComponentDir);
-            });
-            test("should copy files from the template", () => {
-                let files: Array<string> = [];
-
-                function testGeneratedFiles(folderName: string) {
-                    const tempDirContents = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName));
-                    const tempDirContentsWithFileTypes = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName), {
-                        withFileTypes: true
-                    });
-
-                    for (let i = 0, contentLength = tempDirContents.length; i < contentLength; i++) {
-                        if (tempDirContentsWithFileTypes[i].isDirectory()) {
-                            testGeneratedFiles(tempDirContents[i]);
-                        } else {
-                            files.push(
-                                folderName
-                                    ? `${folderName}/${tempDirContents[i]}`
-                                    : tempDirContents[i]
-                            );
-                        }
-                    }
-                }
-                
-                testGeneratedFiles("");
-                expect(files).toEqual(expectedGeneratedComponentTemplateFiles);
-            });
-            test("should be able to run the build", () => {
-                expect(
-                    () => {
-                        execSync(`cd ${tempDir} && npm run build`);
-                    }
-                ).not.toThrow();
-            });
-        });
-        test.describe("badge", () => {
-            test.beforeAll(() => {
-                setup(tempDir, tempComponentDir);
-                execSync(`cd ${tempDir} && npm run fast:init`);
-                setup(tempDir, tempComponentDir);
-                execSync(`cd ${tempDir} && npm run fast:add-foundation-component:badge`);
-            });
-            test.afterAll(() => {
-                teardown(tempDir, tempComponentDir);
-            });
-            test("should copy files from the template", () => {
-                let files: Array<string> = [];
-
-                function testGeneratedFiles(folderName: string) {
-                    const tempDirContents = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName));
-                    const tempDirContentsWithFileTypes = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName), {
-                        withFileTypes: true
-                    });
-
-                    for (let i = 0, contentLength = tempDirContents.length; i < contentLength; i++) {
-                        if (tempDirContentsWithFileTypes[i].isDirectory()) {
-                            testGeneratedFiles(tempDirContents[i]);
-                        } else {
-                            files.push(
-                                folderName
-                                    ? `${folderName}/${tempDirContents[i]}`
-                                    : tempDirContents[i]
-                            );
-                        }
-                    }
-                }
-                
-                testGeneratedFiles("");
-                expect(files).toEqual(expectedGeneratedComponentTemplateFiles);
-            });
-            test("should be able to run the build", () => {
-                expect(
-                    () => {
-                        execSync(`cd ${tempDir} && npm run build`);
-                    }
-                ).not.toThrow();
-            });
-        });
-        test.describe("disclosure", () => {
-            test.beforeAll(() => {
-                setup(tempDir, tempComponentDir);
-                execSync(`cd ${tempDir} && npm run fast:init`);
-                setup(tempDir, tempComponentDir);
-                execSync(`cd ${tempDir} && npm run fast:add-foundation-component:disclosure`);
-            });
-            test.afterAll(() => {
-                teardown(tempDir, tempComponentDir);
-            });
-            test("should copy files from the template", () => {
-                let files: Array<string> = [];
-
-                function testGeneratedFiles(folderName: string) {
-                    const tempDirContents = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName));
-                    const tempDirContentsWithFileTypes = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName), {
-                        withFileTypes: true
-                    });
-
-                    for (let i = 0, contentLength = tempDirContents.length; i < contentLength; i++) {
-                        if (tempDirContentsWithFileTypes[i].isDirectory()) {
-                            testGeneratedFiles(tempDirContents[i]);
-                        } else {
-                            files.push(
-                                folderName
-                                    ? `${folderName}/${tempDirContents[i]}`
-                                    : tempDirContents[i]
-                            );
-                        }
-                    }
-                }
-                
-                testGeneratedFiles("");
-                expect(files).toEqual(expectedGeneratedComponentTemplateFiles);
-            });
-            test("should be able to run the build", () => {
-                expect(
-                    () => {
-                        execSync(`cd ${tempDir} && npm run build`);
-                    }
-                ).not.toThrow();
-            });
-        });
-        test.describe("card", () => {
-            test.beforeAll(() => {
-                setup(tempDir, tempComponentDir);
-                execSync(`cd ${tempDir} && npm run fast:init`);
-                setup(tempDir, tempComponentDir);
-                execSync(`cd ${tempDir} && npm run fast:add-foundation-component:card`);
-            });
-            test.afterAll(() => {
-                teardown(tempDir, tempComponentDir);
-            });
-            test("should copy files from the template", () => {
-                let files: Array<string> = [];
-
-                function testGeneratedFiles(folderName: string) {
-                    const tempDirContents = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName));
-                    const tempDirContentsWithFileTypes = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName), {
-                        withFileTypes: true
-                    });
-
-                    for (let i = 0, contentLength = tempDirContents.length; i < contentLength; i++) {
-                        if (tempDirContentsWithFileTypes[i].isDirectory()) {
-                            testGeneratedFiles(tempDirContents[i]);
-                        } else {
-                            files.push(
-                                folderName
-                                    ? `${folderName}/${tempDirContents[i]}`
-                                    : tempDirContents[i]
-                            );
-                        }
-                    }
-                }
-                
-                testGeneratedFiles("");
-                expect(files).toEqual(expectedGeneratedComponentTemplateFiles);
-            });
-            test("should be able to run the build", () => {
-                expect(
-                    () => {
-                        execSync(`cd ${tempDir} && npm run build`);
-                    }
-                ).not.toThrow();
-            });
-        });
-        test.describe("dialog", () => {
-            test.beforeAll(() => {
-                setup(tempDir, tempComponentDir);
-                execSync(`cd ${tempDir} && npm run fast:init`);
-                setup(tempDir, tempComponentDir);
-                execSync(`cd ${tempDir} && npm run fast:add-foundation-component:dialog`);
-            });
-            test.afterAll(() => {
-                teardown(tempDir, tempComponentDir);
-            });
-            test("should copy files from the template", () => {
-                let files: Array<string> = [];
-
-                function testGeneratedFiles(folderName: string) {
-                    const tempDirContents = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName));
-                    const tempDirContentsWithFileTypes = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName), {
-                        withFileTypes: true
-                    });
-
-                    for (let i = 0, contentLength = tempDirContents.length; i < contentLength; i++) {
-                        if (tempDirContentsWithFileTypes[i].isDirectory()) {
-                            testGeneratedFiles(tempDirContents[i]);
-                        } else {
-                            files.push(
-                                folderName
-                                    ? `${folderName}/${tempDirContents[i]}`
-                                    : tempDirContents[i]
-                            );
-                        }
-                    }
-                }
-                
-                testGeneratedFiles("");
-                expect(files).toEqual(expectedGeneratedComponentTemplateFiles);
-            });
-            test("should be able to run the build", () => {
-                expect(
-                    () => {
-                        execSync(`cd ${tempDir} && npm run build`);
-                    }
-                ).not.toThrow();
-            });
-        });
-        test.describe("number-field", () => {
-            test.beforeAll(() => {
-                setup(tempDir, tempComponentDir);
-                execSync(`cd ${tempDir} && npm run fast:init`);
-                setup(tempDir, tempComponentDir);
-                execSync(`cd ${tempDir} && npm run fast:add-foundation-component:number-field`);
-            });
-            test.afterAll(() => {
-                teardown(tempDir, tempComponentDir);
-            });
-            test("should copy files from the template", () => {
-                let files: Array<string> = [];
-
-                function testGeneratedFiles(folderName: string) {
-                    const tempDirContents = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName));
-                    const tempDirContentsWithFileTypes = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName), {
-                        withFileTypes: true
-                    });
-
-                    for (let i = 0, contentLength = tempDirContents.length; i < contentLength; i++) {
-                        if (tempDirContentsWithFileTypes[i].isDirectory()) {
-                            testGeneratedFiles(tempDirContents[i]);
-                        } else {
-                            files.push(
-                                folderName
-                                    ? `${folderName}/${tempDirContents[i]}`
-                                    : tempDirContents[i]
-                            );
-                        }
-                    }
-                }
-                
-                testGeneratedFiles("");
-                expect(files).toEqual(expectedGeneratedComponentTemplateFiles);
-            });
-            test("should be able to run the build", () => {
-                expect(
-                    () => {
-                        execSync(`cd ${tempDir} && npm run build`);
-                    }
-                ).not.toThrow();
-            });
-        });
-        test.describe("search", () => {
-            test.beforeAll(() => {
-                setup(tempDir, tempComponentDir);
-                execSync(`cd ${tempDir} && npm run fast:init`);
-                setup(tempDir, tempComponentDir);
-                execSync(`cd ${tempDir} && npm run fast:add-foundation-component:search`);
-            });
-            test.afterAll(() => {
-                teardown(tempDir, tempComponentDir);
-            });
-            test("should copy files from the template", () => {
-                let files: Array<string> = [];
-
-                function testGeneratedFiles(folderName: string) {
-                    const tempDirContents = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName));
-                    const tempDirContentsWithFileTypes = fs.readdirSync(path.resolve(tempDir, "src/components/test-component", folderName), {
-                        withFileTypes: true
-                    });
-
-                    for (let i = 0, contentLength = tempDirContents.length; i < contentLength; i++) {
-                        if (tempDirContentsWithFileTypes[i].isDirectory()) {
-                            testGeneratedFiles(tempDirContents[i]);
-                        } else {
-                            files.push(
-                                folderName
-                                    ? `${folderName}/${tempDirContents[i]}`
-                                    : tempDirContents[i]
-                            );
-                        }
-                    }
-                }
-                
-                testGeneratedFiles("");
-                expect(files).toEqual(expectedGeneratedComponentTemplateFiles);
-            });
-            test("should be able to run the build", () => {
-                expect(
-                    () => {
-                        execSync(`cd ${tempDir} && npm run build`);
-                    }
-                ).not.toThrow();
-            });
-        });
-    });
 });
